Convert logger view to more sorted format
