--- Modules/Wider.Tools.Logger/LoggerView.xaml.cs
@@ -1,36 +1,0 @@
-﻿#region License
-
-// Copyright (c) 2013 Chandramouleswaran Ravichandran
-// 
-// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-// 
-// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-// 
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-#endregion
-
-using System;
-using System.ComponentModel;
-using System.Windows.Controls;
-using Wider.Interfaces;
-
-namespace Wider.Tools.Logger
-{
-    /// <summary>
-    /// Interaction logic for LoggerView.xaml
-    /// </summary>
-    internal partial class LoggerView : UserControl, IContentView, INotifyPropertyChanged
-    {
-        public LoggerView() => InitializeComponent();
-
-        /// <summary>
-        /// Should be called when a property value has changed
-        /// </summary>
-        /// <param name="propertyName">The property name</param>
-        protected virtual void RaisePropertyChanged(String propertyName) => 
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-
-        public event PropertyChangedEventHandler PropertyChanged;
-    }
-}
