+++ Modules/Wider.Tools.Logger/Views/LoggerView.xaml
@@ -1,7 +1,11 @@
-﻿<UserControl x:Class="Wider.Tools.Logger.LoggerView" xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" x:ClassModifier="internal"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" mc:Ignorable="d" d:DesignHeight="300" d:DesignWidth="300">
+﻿<UserControl 
+    x:Class="Wider.Tools.Logger.Views.LoggerView" 
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" x:ClassModifier="internal"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+    
+    mc:Ignorable="d" d:DesignHeight="300" d:DesignWidth="300">
     <TextBox Margin="5" Text="{Binding Path=Text, Mode=OneWay}" IsReadOnly="True" BorderBrush="White">
         <TextBox.Style>
             <Style TargetType="TextBox">
