+++ Modules/Wider.Tools.Logger/Views/LoggerView.xaml.cs
@@ -10,22 +10,27 @@
 
 #endregion
 
-using Prism.Mvvm;
 using System;
 using System.ComponentModel;
+using System.Windows.Controls;
 using Wider.Interfaces;
-using Wider.Interfaces.Services;
 
-namespace Wider.Tools.Logger
+namespace Wider.Tools.Logger.Views
 {
-    internal class LoggerModel : BindableBase
+    /// <summary>
+    /// Interaction logic for LoggerView.xaml
+    /// </summary>
+    internal partial class LoggerView : UserControl, IContentView, INotifyPropertyChanged
     {
-        public String Text { get; private set; }
+        public LoggerView() => InitializeComponent();
 
-        public void AddLog(ILoggerService logger)
-        {
-            Text = logger.Message + "\n" + Text;
-            RaisePropertyChanged("Text");
-        }
+        /// <summary>
+        /// Should be called when a property value has changed
+        /// </summary>
+        /// <param name="propertyName">The property name</param>
+        protected virtual void RaisePropertyChanged(String propertyName) => 
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+
+        public event PropertyChangedEventHandler PropertyChanged;
     }
 }
